requirements.txt
huggingface_hub>=0.23.0
flask==3.0.3
python-dotenv 
requests